# frontend/src/types/openapi.yaml
...
openapi: 3.0.1
info:
  title: travelPlan API Test
  description: travelPlan의 api 테스트입니다.
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  "/recruits/{recruitId}":
    get:
      tags:
      - recruit-controller
      operationId: getRecruit
      parameters:
      - name: recruitId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataRecruitDetailResponseDto"
    put:
      tags:
      - recruit-controller
      operationId: updateRecruit
      parameters:
      - name: recruitId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecruitRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataRecruitDetailResponseDto"
    delete:
      tags:
      - recruit-controller
      operationId: deleteRecruit
      parameters:
      - name: recruitId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
  "/place/{id}":
    get:
      tags:
      - Place API
      summary: 여행지 조회
      operationId: getPlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataPlaceResDto"
    put:
      tags:
      - Place API
      summary: 여행지 정보 수정
      operationId: updatePlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlaceUpdateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataPlaceResDto"
    delete:
      tags:
      - Place API
      summary: 여행지 삭제
      operationId: deletePlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
...
  "/recruits":
    get:
      tags:
      - recruit-controller
      operationId: getRecruits
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListRecruitListResponseDto"
    post:
      tags:
      - recruit-controller
      operationId: createRecruit
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecruitRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataRecruitDetailResponseDto"
  "/recruits/{recruitId}/applies":
    get:
      tags:
      - apply-controller
      operationId: getApplies
      parameters:
      - name: recruitId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListApplyResponseDto"
    post:
      tags:
      - apply-controller
      operationId: createApply
      parameters:
      - name: recruitId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ApplyCreateRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataApplyResponseDto"
...
  "/place":
    get:
      tags:
      - Place API
      summary: 전체 여행지 조회
      operationId: getAllPlaces
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListPlaceResDto"
    post:
      tags:
      - Place API
      summary: 여행지 등록
      operationId: createPlace
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlaceCreateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataPlaceResDto"
...
  "/recruits/search":
    get:
      tags:
      - recruit-controller
      operationId: searchRecruits
      parameters:
      - name: keyword
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListRecruitListResponseDto"
  "/recruits/search3":
    get:
      tags:
      - recruit-controller
      operationId: searchAndFilter
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: cityName
        in: query
        required: false
        schema:
          type: string
      - name: isClosed
        in: query
        required: false
        schema:
          type: boolean
      - name: startDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: travelStyle
        in: query
        required: false
        schema:
          type: string
      - name: sameGender
        in: query
        required: false
        schema:
          type: boolean
      - name: sameAge
        in: query
        required: false
        schema:
          type: boolean
      - name: minBudget
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxBudget
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minGroupSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxGroupSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListRecruitListResponseDto"
  "/recruits/search2":
    get:
      tags:
      - recruit-controller
      operationId: findRecruitsByIsClosed
      parameters:
      - name: isClosed
        in: query
        required: true
        schema:
          type: boolean
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListRecruitListResponseDto"
...
  "/api/reviews/place/{placeId}":
    get:
      tags:
      - review-controller
      operationId: getReviewsByPlace
      parameters:
      - name: placeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListReviewResponseDto"
...
  "/recruits/{recruitId}/applies/{applyId}":
    delete:
      tags:
      - apply-controller
      operationId: deleteApply
      parameters:
      - name: applyId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
...
components:
  schemas:
...
    RecruitRequestDto:
      type: object
      properties:
        placeId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelStyle:
          type: string
          enum:
          - SIGHTSEEING
          - RELAXATION
          - ADVENTURE
          - GOURMET
          - SHOPPING
        sameGender:
          type: boolean
        sameAge:
          type: boolean
        budget:
          type: integer
          format: int32
        groupSize:
          type: integer
          format: int32
        closed:
          type: boolean
    RecruitDetailResponseDto:
      type: object
      properties:
        recruitId:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        memberProfileImage:
          type: string
        memberNickname:
          type: string
        placeId:
          type: integer
          format: int64
        placeCityName:
          type: string
        placePlaceName:
          type: string
        title:
          type: string
        content:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelStyle:
          type: string
          enum:
          - SIGHTSEEING
          - RELAXATION
          - ADVENTURE
          - GOURMET
          - SHOPPING
        sameGender:
          type: boolean
        sameAge:
          type: boolean
        budget:
          type: integer
          format: int32
        groupSize:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closed:
          type: boolean
    RsDataRecruitDetailResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/RecruitDetailResponseDto"
    PlaceUpdateReqDto:
      type: object
      properties:
        cityName:
          type: string
        placeName:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
          - PLACE
          - STAY
          - RESTAURANT
          - CAFE
          - ETC
    PlaceResDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cityName:
          type: string
        placeName:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
          - PLACE
          - STAY
          - RESTAURANT
          - CAFE
          - ETC
    RsDataPlaceResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/PlaceResDto"
...
    ReviewResponseDto:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        rating:
          type: number
          format: double
        memberId:
          type: integer
          format: int64
        memberName:
          type: string
        placeId:
          type: integer
          format: int64
        viewCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        commentCount:
          type: integer
          format: int32
...
    ApplyCreateRequestDto:
      type: object
      properties:
        content:
          type: string
    ApplyResponseDto:
      type: object
      properties:
        applyId:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        memberProfileImage:
          type: string
        memberNickname:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RsDataApplyResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/ApplyResponseDto"
...
    PlaceCreateReqDto:
      required:
      - category
      - cityName
      - placeName
      type: object
      properties:
        cityName:
          type: string
        placeName:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
          - PLACE
          - STAY
          - RESTAURANT
          - CAFE
          - ETC
...
    RecruitListResponseDto:
      type: object
      properties:
        recruitId:
          type: integer
          format: int64
        memberProfileImage:
          type: string
        memberNickname:
          type: string
        placeCityName:
          type: string
        placePlaceName:
          type: string
        title:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelStyle:
          type: string
          enum:
          - SIGHTSEEING
          - RELAXATION
          - ADVENTURE
          - GOURMET
          - SHOPPING
        sameGender:
          type: boolean
        sameAge:
          type: boolean
        budget:
          type: integer
          format: int32
        groupSize:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closed:
          type: boolean
    RsDataListRecruitListResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/RecruitListResponseDto"
    RsDataListApplyResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/ApplyResponseDto"
    RsDataListPlaceResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/PlaceResDto"
    RsDataListReviewResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/ReviewResponseDto"
...
    RsDataVoid:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: object
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT